Broke down UserService into AuthService and UserService
