+++ backend/src/main/java/com/htecgroup/skynest/controller/RegistrationController.java
@@ -4,6 +4,7 @@ import com.htecgroup.skynest.model.request.UserPasswordResetRequest;
 import com.htecgroup.skynest.model.request.UserRegisterRequest;
 import com.htecgroup.skynest.model.response.ErrorMessage;
 import com.htecgroup.skynest.model.response.UserResponse;
+import com.htecgroup.skynest.service.AuthService;
 import com.htecgroup.skynest.service.UserService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
@@ -28,7 +29,9 @@ import static com.htecgroup.skynest.util.UrlUtil.*;
 @Log4j2
 @Tag(name = "Registration API", description = "Operations related to user registration")
 public class RegistrationController {
+
   private UserService userService;
+  private AuthService authService;
 
   @Operation(summary = "Register a new user")
   @ApiResponses(
@@ -117,6 +120,7 @@ public class RegistrationController {
 
     ResponseEntity<UserResponse> responseEntity =
         new ResponseEntity<>(userService.registerUser(userRegisterRequest), HttpStatus.OK);
+    authService.sendVerificationEmail(userRegisterRequest.getEmail());
     return responseEntity;
   }
 
@@ -148,7 +152,7 @@ public class RegistrationController {
       })
   @PostMapping(CONFIRM_EMAIL_URL)
   public ResponseEntity<String> confirmEmail(@RequestParam String token) {
-    String response = userService.confirmEmail(token);
+    String response = authService.confirmEmail(token);
     log.info(response);
     return ResponseEntity.ok(response);
   }
@@ -178,7 +182,7 @@ public class RegistrationController {
       })
   @PostMapping(RESEND_EMAIL_URL)
   public ResponseEntity<String> resendUserEmail(@RequestParam String email) {
-    userService.sendVerificationEmail(email);
+    authService.sendVerificationEmail(email);
     String response = "Email resent successfully";
     log.info(response);
     return ResponseEntity.ok(response);
@@ -226,7 +230,7 @@ public class RegistrationController {
       })
   @PostMapping(PASSWORD_RESET_URL)
   public ResponseEntity<String> requestPasswordReset(@RequestParam String email) {
-    userService.sendPasswordResetEmail(email);
+    authService.sendPasswordResetEmail(email);
     String response = "Password reset email sent";
     log.info(response);
     return ResponseEntity.ok(response);
@@ -266,7 +270,7 @@ public class RegistrationController {
   @PutMapping(PASSWORD_RESET_URL)
   public ResponseEntity<String> confirmPasswordReset(
       @Valid @RequestBody UserPasswordResetRequest userPasswordResetRequest) {
-    String response = userService.resetPassword(userPasswordResetRequest);
+    String response = authService.resetPassword(userPasswordResetRequest);
     log.info(response);
     return ResponseEntity.ok(response);
   }
