+++ backend/src/main/java/com/htecgroup/skynest/security/WebSecurity.java
@@ -3,9 +3,9 @@ package com.htecgroup.skynest.security;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.htecgroup.skynest.filter.CustomAuthenticationFilter;
 import com.htecgroup.skynest.filter.CustomAuthorizationFilter;
+import com.htecgroup.skynest.service.AuthService;
 import com.htecgroup.skynest.service.InvalidJwtService;
 import com.htecgroup.skynest.service.LoginAttemptService;
-import com.htecgroup.skynest.service.UserService;
 import com.htecgroup.skynest.util.UrlUtil;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Configuration;
@@ -27,7 +27,7 @@ public class WebSecurity extends WebSecurityConfigurerAdapter {
   private final CustomUserDetailsService userDetailsService;
   private final BCryptPasswordEncoder bCryptPasswordEncoder;
   private final ObjectMapper objectMapper;
-  private final UserService userService;
+  private final AuthService authService;
   private final LoginAttemptService loginAttemptService;
   private final InvalidJwtService invalidJwtService;
 
@@ -54,7 +54,7 @@ public class WebSecurity extends WebSecurityConfigurerAdapter {
         new CustomAuthenticationFilter(
             authenticationManager(),
             userDetailsService,
-            userService,
+            authService,
             objectMapper,
             loginAttemptService);
     filter.setFilterProcessesUrl(
