+++ backend/src/main/java/com/htecgroup/skynest/service/UserService.java
@@ -2,7 +2,6 @@ package com.htecgroup.skynest.service;
 
 import com.htecgroup.skynest.model.dto.UserDto;
 import com.htecgroup.skynest.model.request.UserEditRequest;
-import com.htecgroup.skynest.model.request.UserPasswordResetRequest;
 import com.htecgroup.skynest.model.request.UserRegisterRequest;
 import com.htecgroup.skynest.model.response.UserResponse;
 
@@ -16,18 +15,6 @@ public interface UserService {
   UserDto findUserByEmail(String email);
 
   List<UserResponse> listAllUsers();
-
-  String confirmEmail(String token);
-
-  UserDto verifyUser(UserDto userDto);
-
-  void sendVerificationEmail(String email);
-
-  void sendPasswordResetEmail(String email);
-
-  String resetPassword(UserPasswordResetRequest userPasswordResetRequest);
-
-  boolean isActive(String email);
 
   void deleteUser(UUID uuid);
